requirements.txt
streamlit==1.32.2
streamlit-option-menu==0.3.13
opencv-python-headless
pillow
numpy
pandas
matplotlib
reportlab
ultralytics
av
streamlit-webrtc
qrcode
openai
streamlit-chat
streamlit-mic-recorder
SpeechRecognition
pydub
gTTS
